Separating the analysis into a separate function
